+++ App.js
@@ -1,25 +1,14 @@
 import { StatusBar } from "expo-status-bar";
 import { useState } from "react";
-import {
-	Button,
-	FlatList,
-	StyleSheet,
-	Text,
-	TextInput,
-	View,
-	ScrollView,
-} from "react-native";
+import { FlatList, StyleSheet, View } from "react-native";
+
+import GoalItem from "./components/GoalItem";
+import GoalInput from "./components/GoalInput";
 
 export default function App() {
-	const [enteredGoalText, setEnterGoalText] = useState("");
 	const [courseGoals, setCourseGoals] = useState([]);
 
-	const goalInputHandler = (enteredText) => {
-		setEnterGoalText(enteredText);
-	};
-
-	const addGoalHandler = () => {
-		setEnterGoalText("");
+	const addGoalHandler = (enteredGoalText) => {
 		setCourseGoals((prev) => [
 			...prev,
 			{ text: enteredGoalText, id: Math.random().toString() },
@@ -28,24 +17,12 @@ export default function App() {
 
 	return (
 		<View style={styles.appContainer}>
-			<View style={styles.inputContainer}>
-				<TextInput
-					style={styles.textInput}
-					placeholder="Your course goal!"
-					onChangeText={goalInputHandler}
-					value={enteredGoalText}
-				/>
-				<Button title="Add Goal" onPress={addGoalHandler} />
-			</View>
+			<GoalInput addGoal={addGoalHandler} />
 			<View style={styles.goalsContainer}>
 				<FlatList
 					data={courseGoals}
 					renderItem={(itemData) => (
-						<View style={styles.goalItem}>
-							<Text style={styles.goalText}>
-								{itemData.item.text}
-							</Text>
-						</View>
+						<GoalItem text={itemData.item.text} />
 					)}
 					keyExtractor={(item, index) => item.id}
 				/>
@@ -60,31 +37,11 @@ const styles = StyleSheet.create({
 		paddingTop: 60,
 		paddingHorizontal: 16,
 	},
-	inputContainer: {
-		flex: 1,
-		flexDirection: "row",
-		justifyContent: "space-between",
-		alignItems: "center",
-		marginBottom: 24,
-		borderBottomWidth: 1,
-		borderColor: "grey",
-	},
-	textInput: {
-		borderWidth: 1,
-		borderColor: "grey",
-		width: "70%",
-		marginRight: 8,
-		padding: 8,
-	},
+
 	goalsContainer: {
 		flex: 5,
 	},
-	goalItem: {
-		margin: 8,
-		padding: 8,
-		borderRadius: 10,
-		backgroundColor: "#1e90ff",
-	},
+
 	goalText: {
 		color: "white",
 	},
